fix: #70 - fixes quick flash on new greeting UI when visited throguh url directly
