+++ src/components/CheckIcon/CheckIcon.js
@@ -1,0 +1,20 @@
+const CheckIcon = () => {
+  return (
+    <div className="flex items-center justify-center">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        className="h-4 w-4 text-green-500"
+        viewBox="0 0 20 20"
+        fill="currentColor"
+      >
+        <path
+          fillRule="evenodd"
+          d="M18.293 3.293a1 1 0 0 1 1.414 1.414l-11 11a1 1 0 0 1-1.414 0l-5-5a1 1 0 0 1 1.414-1.414l4.293 4.293 10-10z"
+          clipRule="evenodd"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export default CheckIcon;
