+++ src/styles/NewUserGreeting.scss
@@ -1,0 +1,3 @@
+.NewUserGreeting {
+  background-color: #013281;
+}
