+++ src/styles/globals.scss
@@ -3,6 +3,7 @@
 @tailwind utilities;
 
 @import "./GetStartedPage.scss";
+@import "./NewUserGreeting.scss";
 
 :root {
 }
